+++ argilla-frontend/CHANGELOG.md
@@ -20,7 +20,7 @@ These are the section headers that we use:
 
 ### Added
 
-- Added `Required/Optional` label on `Field dataset settings tab` and `Question dataset settings tab` ([#5394](https://github.com/argilla-io/argilla/pull/5394))
+- Added `Required/Optional` label on `Field dataset settings tab` and `Question dataset settings tab`. ([#5394](https://github.com/argilla-io/argilla/pull/5394))
 - Added new `ChatField`. ([#5376](https://github.com/argilla-io/argilla/pull/5376))
 
 ## [2.1.0](https://github.com/argilla-io/argilla/compare/v2.0.1...v2.1.0)